_damon_result/record_to_damon_result: Return error instead of exit()

+++ damo_record.py
@@ -31,7 +31,10 @@ data_for_cleanup = DataForCleanup()
 def change_rfile_format(rfile_path, src_format, dst_format, dst_permission):
     rfile_path_mid = rfile_path + '.mid'
     os.rename(rfile_path, rfile_path_mid)
-    result = _damon_result.parse_damon_result(rfile_path_mid, src_format)
+    result, err = _damon_result.parse_damon_result(rfile_path_mid, src_format)
+    if err:
+        raise Exception('parsing DAMON output failed (%s)' % err)
+
     _damon_result.write_damon_result(result, rfile_path, dst_format,
             dst_permission)
     os.remove(rfile_path_mid)
